relatorio finalizado

+++ AvivDatabasePopulator/Carthage/Checkouts/swift-sdk/Scripts/travis/publish-documentation.sh
@@ -3,25 +3,32 @@
 
 set -e
 
-gem install jazzy --silent
+if [ "$is_travis" = "TRAVIS" ]; then
+    gem install jazzy
 
-# Configure Travis to be able to push to the Github repo
-git config --global user.email "travis@travis-ci.org"
-git config --global user.name "Travis CI"
-git config --replace-all remote.origin.fetch +refs/heads/*:refs/remotes/origin/*
-git remote rm origin
-git remote add origin https://watson-developer-cloud:${GH_TOKEN}@github.com/watson-developer-cloud/swift-sdk.git
+    # Configure Travis to be able to push to the Github repo
+    git config --global user.email "travis@travis-ci.org"
+    git config --global user.name "Travis CI"
+    git config --replace-all remote.origin.fetch +refs/heads/*:refs/remotes/origin/*
+    git remote rm origin
+    git remote add origin https://watson-developer-cloud:${GH_TOKEN}@github.com/watson-developer-cloud/swift-sdk.git
+fi
+
 git fetch
 git checkout master
 latestVersion=$(git describe --abbrev=0 --tags)
 
-# Generate the API docs
-./Scripts/generate-documentation.sh
+git clone --quiet --branch=gh-pages git@github.com:watson-developer-cloud/swift-sdk.git gh-pages > /dev/null
 
-# Push newly-generated docs to the gh-pages branch
-git checkout --track origin/gh-pages
-cp -r gh-pages/* .
-rm -rf gh-pages/
-git add .
-git commit -m "SDK docs for release ${latestVersion}"
-git push --set-upstream origin gh-pages
+# Delete all the old docs (but not the docs directory -- this is hand written)
+(cd gh-pages && git rm -rf css img index.html js services undocumented.json)
+
+# Generate the API docs
+./Scripts/generate-documentation.sh gh-pages
+
+# Commit and push the newly generated API docs
+pushd gh-pages
+  git add .
+  git commit -m "SDK docs for release ${latestVersion}"
+  git push
+popd
